[test] Get rid of EnergyPlus v8.8 dep in all tests

+++ tests/testthat/test-run.R
@@ -614,27 +614,31 @@ test_that("energyplus()", {
     unlink(c(path_exp, out_dir), recursive = TRUE)
 
     if(Sys.getenv("_EPLUSR_SKIP_TESTS_BASEMENT_") == "") {
-        # can run Basement
-        path_base <- copy_eplus_example(LATEST_EPLUS_VER, "LgOffVAVusingBasement.idf")
-        expect_error(energyplus(path_base, NULL, out_dir, echo = FALSE))
-        # modify the input in order to reduce the simulation time
-        l <- read_lines(path_base)
-        l[grepl("IYRS: Maximum number of yearly iterations", string), string := "1;"]
-        write_lines(l, path_base)
-        res_base <- energyplus(path_base, weather, out_dir, design_day = TRUE, echo = FALSE)
-        expect_equal(length(res_base$file), 57L)
-        # NOTE: From EnergyPlus v22.1.0, Sqlite.err is only generated if there is
-        # a SQLite output
-        expect_equal({files <- unlist(res_base$file); files <- files[!is.na(files)]; length(files)}, 26L)
-        expect_equal(sum(!file.exists(file.path(out_dir, files))), 0L)
-        expect_equal(sort(unname(files[names(files) != "epw"])), list.files(out_dir, "LgOffVAVusingBasement"))
-        expect_equal(res_base$run[program == "Basement", exit_status], list(0L))
-        expect_true(file.exists(file.path(out_dir, files["expidf"])))
-        expect_true(file.exists(file.path(out_dir, files["bsmt_idf"])))
-        expect_true(file.exists(file.path(out_dir, files["bsmt_csv"])))
-        expect_true(file.exists(file.path(out_dir, files["bsmt_out"])))
-        expect_true(file.exists(file.path(out_dir, files["bsmt_audit"])))
-        unlink(c(path_base, out_dir), recursive = TRUE)
+        # NOTE: There is a bug in the basement preprocessor in EnergyPlus from v9.4 to v22.1
+        # Ref: https://github.com/NREL/EnergyPlus/pull/9356
+        if (numeric_version(LATEST_EPLUS_VER) > 22.1) {
+            # can run Basement
+            path_base <- copy_eplus_example(LATEST_EPLUS_VER, "LgOffVAVusingBasement.idf")
+            expect_error(energyplus(path_base, NULL, out_dir, echo = FALSE))
+            # modify the input in order to reduce the simulation time
+            l <- read_lines(path_base)
+            l[grepl("IYRS: Maximum number of yearly iterations", string), string := "1;"]
+            write_lines(l, path_base)
+            res_base <- energyplus(path_base, weather, out_dir, design_day = TRUE, echo = FALSE)
+            expect_equal(length(res_base$file), 57L)
+            # NOTE: From EnergyPlus v22.1.0, Sqlite.err is only generated if there is
+            # a SQLite output
+            expect_equal({files <- unlist(res_base$file); files <- files[!is.na(files)]; length(files)}, 26L)
+            expect_equal(sum(!file.exists(file.path(out_dir, files))), 0L)
+            expect_equal(sort(unname(files[names(files) != "epw"])), list.files(out_dir, "LgOffVAVusingBasement"))
+            expect_equal(res_base$run[program == "Basement", exit_status], list(0L))
+            expect_true(file.exists(file.path(out_dir, files["expidf"])))
+            expect_true(file.exists(file.path(out_dir, files["bsmt_idf"])))
+            expect_true(file.exists(file.path(out_dir, files["bsmt_csv"])))
+            expect_true(file.exists(file.path(out_dir, files["bsmt_out"])))
+            expect_true(file.exists(file.path(out_dir, files["bsmt_audit"])))
+            unlink(c(path_base, out_dir), recursive = TRUE)
+        }
 
         # can run Slab
         path_slab <- copy_eplus_example(LATEST_EPLUS_VER, "5ZoneAirCooledWithSlab.idf")
